Reworked Docker utils so it deals with cases where a connection to Docker fails.
The Docker utils lib typically connects to the Docker daemon to get info about
containers.

In some cases this connection can fail. In particular, when Docker containers
are configured with "--restart" and a machine is rebooted, Docker will try to
restart the containers but will not respond via it's API until the containers
are restarted. See sysbox issue #184.

To ensure Sysbox works properly in such cases, this change modifies the Docker
utils so that it can deal with scenarios where Docker is not responding via it's
API. In particular, function ContainerIsDocker() was reworked so that it can
tell if a container is a Docker container without necessarily querying this info
from Docker itself.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>