Prevent golang deadlock detection in runIDMapMountCheckOnHost().
Small change to prevent golang runtime from getting confused into thinking that
there is a deadlock in runIDMapMountCheckOnHost().

Without this change, Golang 1.19+ was sporadically complaining as follows:

"Starting sysbox-mgr
fatal error: all goroutines are asleep - deadlock!"

This occurs because runIDMapMountCheckOnHost() forks a process into a user-ns
that sleeps forver, and then kills the process and waits for it to exit. During
this wait, there is a window of time where all goroutines are waiting (though
not forever), and this causes the golang runtime to get confused into
thinking there is a deadlock.

This commit fixes this.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>