idShiftUtils: add CheckACLSupport() function.
Turns out not all filesystems in Linux support Access Control Lists (ACL).
For example, support for ACL can be enabled or disabled in ZFS.

This commit adds a function to check if ACL is supported on a given
filesystem. If not supported, then function ShiftIdsWithChown() will not attempt
to change ownserhip attributes on the ACL, as otherwise it would received an
error.

Fixes Sysbox issue 849 (https://github.com/nestybox/sysbox/issues/849).

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>