Fix bug in shiftfs checker.
There's a bug in the shiftfs checker that is causing it report false negatives
(i.e., it believes shiftfs is not working in hosts where it actually is
working).

The bug was that shiftfs checker was creating a test dir where shiftfs will be
mounted, and chowning that dir to the user-ns mapped uid:gid. However that dir
must not be chowned, since the shiftfs mount on it already ensures that root in
the user-ns will be able to access it. In fact by chowning it we are causing
shiftfs to not work properly on it.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>