Improvement in rootfs cloner chown of hardlinks.
Improve the way the rootfs cloner chowns hardlinked files to the container's
user-ns user-ID range. Instead of chowning the first such link and assume the
rest will chown (because they are supposed to be hardlinked), explicitly chown
all the hardlinked files. This is needed because when the rootfs cloner copies
the container's rootfs to /var/lib/sysbox/rootfs/<uuid>, the hardlinking between
rootfs files that may have been present in the image is not kept, so chowing one
file no longer chowns files that are in theory hardlinked to it.

This change is actually a short-term work-around for a deeper problem: keeping
the hardlinks in the original container image when doing rootfs cloning.

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>